refactor: fix clippy warnings in cache, http, lightpanda, parsers
Dead code annotations, unnecessary casts, map_or→is_some_and, collapsible if let, clamp patterns, filter_map→map simplifications.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src/http.rs
@@ -117,7 +117,7 @@ impl HttpClient {
 
         // Retry loop with exponential backoff
         let mut last_err: Option<anyhow::Error> = None;
-        let max_retries = self.settings.retries as u32;
+        let max_retries = self.settings.retries;
 
         for attempt in 0..=max_retries {
             if attempt > 0 {
